requirements.txt
# Python package dependencies for DXFPipeline
# Install with: pip install -r requirements.txt
#
# External binary: Inkscape (used by rasterise_tiles.py for SVG→PNG rasterisation)
#   Windows : https://inkscape.org/release/  (add to PATH or pass --inkscape)
#   Linux   : sudo apt install inkscape
#   macOS   : brew install inkscape

ezdxf>=1.3
Pillow>=10.0
